fix: reorganizando os arquivos da api, levando para os diretórios certos (pasta src e backend estavam duplicadas)
